Made tooltip text easier to interpret: - removed day out of 366 - reduced numbers to 2 decimal places - clarified that growth rate is "per day" (is this correct?)
Removed circles - I don't think they work visually. We'll need to consider
some other way of communicating error if that's important.

Also, I think errors should be communicated as a pct... right now
I'm guessing they're in units of cases/million? But the growth rate
is not in cases/million, and people will see this as an error associated
with the growth rate...