Feedback Updates
USER INTERFACE

(Solved) The slower/faster buttons should only change the speed. If the animation is not playing, it shouldn't start playing when the user presses "slower" or "faster".
(Solved) The speed should be indicated as soon as the page loads. Right now the circles are all gray on page load.
(Solved) Do we really need a reset button? It is so easy to pick a speed now, it seems the reset button isn't necessary.
(Solved) Can you try adding tooltips for each button, with just one word (e.g. "play", "slower"...). I think it is really easy to do, I haven't done it before but check this thread: https://stackoverflow.com/questions/2238239/tooltips-for-button

CODE
(Working...) The code can still be reduced much more by using for loops instead of if statements. But we can discuss this... for now the user interface is most important.